Add a new subclass of ActiveRecord::Relation for CPK
This adds a subclass to avoid problems with inline caches when
dealing with Relation objects that need CPK behavior. It's not
exactly pretty because it's still operating much like a monkey
patch. The removal of the singleton class codepath should also
lighten memory usage slightly.

To work, the relation behavior selection is now elevated to
`ActiveRecord::Relation.new`. This will return an object based on
the model class passed in. `new` is also overwritten for the
subclass since we want to avoid an infinite loop. An alternative
would be to check self in the other new and simply call the
original new in that case.

In rough trials, some problematic code on implementations like
JRuby went from taking hours to complete to roughly ten minutes.
I'd consider adding benchmarks to keep this in check but at least
we can confirm that this moves performance in the right
direction.